fix(0.6.1): drive runtime global-frame tick from monotonic world tick
Symptom: with Gem style, casting slot 1 (Lapidary) locked out slots 2/3
(Fracture/Refraction) forever - every later cast returned "Lapidary is
still casting for 0.3s", so all controls appeared to "repeat lapidary".
Affected every style, not just Gem.

Root cause: MotmServerTickSystem.tick(float, int tick, Store) passed the
ECS `tick` int into onServerTick -> MotmRuntimeLoop.tick -> claimGlobalFrame.
On 0.6.1 that `tick` int is no longer globally monotonic, so
claimGlobalFrame(serverTick) returned false after the first frame and the
per-frame global ticks stopped running: tickStatusEffects,
tickElementalReactions, tickStyleCooldowns, tickResources. Action windows
therefore never advanced (isFinished never true), so isActionLocked stayed
true permanently and blocked all subsequent casts. Per-world work outside
the guard (field pulses, HUD) kept running, which is why the field visuals
still ticked while cooldowns/windows froze.

Fix: pass the monotonic world tick (store.getExternalData().getWorld()
.getTick(), already computed for per-store dedup) instead of the ECS int,
and widen MenteesMod.onServerTick to long.

Verified in-world on 0.6.1: Lapidary/Fracture/Refraction each cast.begin +
cast.end with distinct effects (persistent gem / 60-dmg burst / 4s buff
aura), zero "still casting" blocks, HUD shows Refraction toggle ON.
compileJava + test + checkArchitecture green.

+++ src/main/java/com/motm/MenteesMod.java
@@ -2196,7 +2196,7 @@ public class MenteesMod extends JavaPlugin {
      * This keeps the runtime systems advancing even before we wire the actual
      * Hytale damage/entity APIs.
      */
-    public void onServerTick(Store<EntityStore> currentStore, int serverTick) {
+    public void onServerTick(Store<EntityStore> currentStore, long serverTick) {
         runtimeLoop.tick(currentStore, serverTick);
     }
